requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# LLM Provider SDKs
google-generativeai>=0.3.0
groq>=0.4.0

# Optional for enhanced features
textstat>=0.7.3  # For readability scoring
